requirements.txt
imbalanced_learn==0.12.2
joblib==1.2.0
jpegio==0.2.8
numpy==1.23.5
numpy==1.24.2
Pillow==9.4.0
Pillow==10.3.0
scikit_learn==1.3.2
scipy==1.9.1
torch==1.10.1+cu102
torchvision==0.11.2+cu102
